+++ lib/src/decode/arm/instruction/data/ldm.dart
@@ -59,13 +59,19 @@ part of '../../instruction.dart';
 /// > of them are for use in conjunction with exception processing, and are not
 /// > described here.
 class LDM extends ArmInstruction {
+  /// Pre/Post indexing bit.
   final int p;
 
+  /// Up/Down bit.
   final int u;
 
+  /// Write-back bit.
   final int w;
 
-  final int sourceRegister;
+  /// PSR & force user bit.
+  final int s;
+
+  final int baseRegister;
 
   final int registerList;
 
@@ -74,7 +80,8 @@ class LDM extends ArmInstruction {
     @required this.p,
     @required this.u,
     @required this.w,
-    @required this.sourceRegister,
+    @required this.s,
+    @required this.baseRegister,
     @required this.registerList,
   }) : super._(condition);
