+++ lib/src/decode/thumb/printer.dart
@@ -1,62 +1,15 @@
-import 'package:binary/binary.dart';
-
+import 'package:armv4t/src/decode/common.dart' as common;
 import 'package:armv4t/src/decode/thumb/instruction.dart';
 import 'package:meta/meta.dart';
 
 class ThumbInstructionPrinter implements ThumbInstructionVisitor<String, void> {
-  /// Given an 8-bit list, where the k-th byte set represents the k-th register.
-  ///
-  /// Any registers in a row are emitted using the format `Rn-Rm`; otherwise
-  /// they are emitted using `Rn`. Every register or register set is separated
-  /// by commas before emitting.
-  ///
-  /// It is considered valid to have _no_ registers (e.g. just the PC/LR).
   @visibleForTesting
   static String describeRegisterList(int registerList, [String suffix]) {
-    if (registerList == 0) {
-      return '';
-    }
-    final output = <String>[];
-    int rangeStart;
-
-    void write(int current) {
-      if (current - 1 == rangeStart) {
-        // Print the previous set bit.
-        output.add('R${rangeStart}');
-      } else {
-        // Print a range beteween the start and the current index.
-        output.add('R${rangeStart}-R${current - 1}');
-      }
-
-      // Clear the range.
-      rangeStart = null;
-    }
-
-    for (var i = 0; i < 8; i++) {
-      if (registerList.isSet(i)) {
-        if (rangeStart == null) {
-          // If we haven't started a range, start one.
-          rangeStart = i;
-        } else {
-          // Otherwise it's fine - we are just waiting until we find a 0 bit.
-        }
-      } else {
-        if (rangeStart != null) {
-          // Write a range from the start register to the previous register.
-          write(i);
-        }
-      }
-    }
-
-    if (rangeStart != null) {
-      write(8);
-    }
-
-    if (suffix != null) {
-      output.add(suffix);
-    }
-
-    return output.join(',');
+    return common.describeRegisterList(
+      registerList,
+      length: 8,
+      suffix: suffix,
+    );
   }
 
   const ThumbInstructionPrinter();
